// mess-web/src/types/razorpay.d.ts
export declare global {
  interface Window {
    Razorpay: new (options: RazorpayOptions) => RazorpayInstance;
  }
}

export interface RazorpayOptions {
  key: string | undefined;
  amount: number;
  currency: string;
  name: string;
  description: string;
  order_id: string;
  handler: (response: RazorpayResponse) => Promise<void>;
  prefill: { name: string; contact: string };
  theme: { color: string };
}

export interface RazorpayInstance {
  open: () => void;
}